commit aula5

+++ aula06/exercicios/exercicio01/Proprietario.java
@@ -1,7 +1,7 @@
 
 public class Proprietario {
     private String nome;
-    private double telefone;
+    private float telefone;
 
     public void exibir() {
         System.out.println(nome + ", " + telefone);
